Move 'app-util-xdm' module to 'com.adobe.marketing.mobile.edge.testapp.xdm.commerce' namespace.
