+++ code/app-util-xdm/src/main/java/com/adobe/marketing/mobile/edge/testapp/xdm/commerce/ProductListAdds.java
@@ -9,24 +9,24 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.tester.xdm.commerce;
+package com.adobe.marketing.mobile.edge.testapp.xdm.commerce;
 
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Class {@code ProductListViews}
- * View or views of a product-list has occurred.
+ * Class {@code ProductListAdds}
+ * Addition of a product to the product list, for example a product is added to a shopping cart.
  *
- * XDM Property Java Object Generated 2020-10-01 15:22:47.697368 -0700 PDT m=+1.799241143 by XDMTool
+ * XDM Property Java Object Generated 2020-10-01 15:22:47.697165 -0700 PDT m=+1.799037386 by XDMTool
  */
 @SuppressWarnings("unused")
-public class ProductListViews implements com.adobe.marketing.mobile.xdm.Property {
+public class ProductListAdds implements com.adobe.marketing.mobile.xdm.Property {
 
 	private String id;
 	private double value;
 
-	public ProductListViews() {}
+	public ProductListAdds() {}
 
 	@Override
 	public Map<String, Object> serializeToXdm() {
