+++ code/app/src/main/java/com/adobe/marketing/mobile/edge/testapp/java/MainActivity.java
@@ -39,8 +39,12 @@ import com.adobe.marketing.mobile.edge.identity.AuthenticatedState;
 import com.adobe.marketing.mobile.edge.identity.Identity;
 import com.adobe.marketing.mobile.edge.identity.IdentityItem;
 import com.adobe.marketing.mobile.edge.identity.IdentityMap;
+import com.adobe.marketing.mobile.edge.testapp.xdm.commerce.Commerce;
+import com.adobe.marketing.mobile.edge.testapp.xdm.commerce.MobileSDKCommerceSchema;
+import com.adobe.marketing.mobile.edge.testapp.xdm.commerce.Order;
+import com.adobe.marketing.mobile.edge.testapp.xdm.commerce.ProductListAdds;
+import com.adobe.marketing.mobile.edge.testapp.xdm.commerce.Purchases;
 import com.adobe.marketing.mobile.services.Log;
-import com.adobe.marketing.tester.xdm.commerce.*;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import java.util.ArrayList;
